requirements.txt
Flask==3.0.1
tensorflow==2.14.0
numpy==1.26.3
Pillow==10.2.0
